added validation of inputs
