+++ apps/funding-platform/package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "ethers": "^5.0.0",
     "express": "^4.18.2",
+    "express-validator": "^6.14.2",
     "lowdb": "^5.0.5",
     "pg-promise": "^10.15.4",
     "rpch-common": "*"
