+++ apps/funding-platform/src/entry-server/index.ts
@@ -1,77 +1,12 @@
-import express, { NextFunction, Request, Response } from "express";
+import express from "express";
 import { AccessTokenService } from "../access-token";
 import { getBalanceForAllChains, getProviders } from "../blockchain";
 import { RequestService } from "../request";
-import { isExpired, smartContractAddresses, validChainIds } from "../utils";
+import { smartContractAddresses, validChainIds } from "../utils";
+import { tokenIsValid } from "./middleware";
+import { body, validationResult, param } from "express-validator";
 
 const app = express();
-
-/**
- * Middleware used to check if token has expired or has been used with too many requests
- * @param accessTokenService
- * @param requestService
- * @param maxAmountOfTokens
- * @param requestFunds
- */
-const tokenIsValid =
-  (
-    accessTokenService: AccessTokenService,
-    requestService: RequestService,
-    maxAmountOfTokens: number,
-    requestFunds?: boolean
-  ) =>
-  async (req: Request, res: Response, next: NextFunction) => {
-    const accessTokenHash = req.headers["x-access-token"] as string;
-    if (!accessTokenHash) return res.status(400).json("Missing Access Token");
-    const dbToken = await accessTokenService.getAccessToken(accessTokenHash);
-    if (!dbToken) return res.status(404).json("Access Token does not exist");
-
-    if (isExpired(dbToken.expired_at)) {
-      return res.status(401).json("Access Token expired");
-    }
-
-    const hasEnough = await doesAccessTokenHaveEnoughBalance({
-      requestService,
-      maxAmountOfTokens,
-      token: dbToken.token,
-      requestAmount: requestFunds ? Number(req.body.amount) : 0,
-    });
-
-    if (!hasEnough) {
-      return res.status(401).json("Exceeded max amount of tokens redeemed");
-    }
-
-    next();
-  };
-
-/**
- * Checks if token can make another request without exceeding max amount of tokens
- * @returns boolean
- */
-export const doesAccessTokenHaveEnoughBalance = async (params: {
-  requestService: RequestService;
-  token: string;
-  maxAmountOfTokens: number;
-  requestAmount?: number;
-}): Promise<Boolean> => {
-  const requestsByAccessToken =
-    await params.requestService.getRequestsByAccessToken(params.token);
-  const totalRequests = requestsByAccessToken?.filter(
-    (req) =>
-      req.status !== "FAILED" &&
-      req.status !== "FAILED-DURING-PROCESSING" &&
-      req.status !== "REJECTED-DURING-PROCESSING"
-  );
-  const sumOfTokensTotalPossibleRequests =
-    totalRequests?.reduce((prev, next) => prev + Number(next.amount), 0) ?? 0;
-
-  const tokenBalanceWithRequestAmount =
-    sumOfTokensTotalPossibleRequests + (params.requestAmount ?? 0);
-  if (params.maxAmountOfTokens < tokenBalanceWithRequestAmount) {
-    return false;
-  }
-  return true;
-};
 
 /**
  * Express server that holds all routes
@@ -105,6 +40,8 @@ export const entryServer = (ops: {
 
   app.post(
     "/api/request/funds/:blockchainAddress",
+    body("amount").notEmpty().bail().isNumeric({ no_symbols: true }),
+    body("chainId").notEmpty().bail().isNumeric(),
     tokenIsValid(
       ops.accessTokenService,
       ops.requestService,
@@ -112,6 +49,11 @@ export const entryServer = (ops: {
       true
     ),
     async (req, res) => {
+      // check if validation failed
+      const errors = validationResult(req);
+      if (!errors.isEmpty()) {
+        return res.status(400).json({ errors: errors.array() });
+      }
       const nodeAddress = String(req.params.blockchainAddress);
       const amount = String(req.body.amount);
       const chainId = Number(req.body.chainId);
@@ -151,12 +93,18 @@ export const entryServer = (ops: {
 
   app.get(
     "/api/request/status/:requestId",
+    param("requestId").isNumeric(),
     tokenIsValid(
       ops.accessTokenService,
       ops.requestService,
       ops.maxAmountOfTokens
     ),
     async (req, res) => {
+      // check if validation failed
+      const errors = validationResult(req);
+      if (!errors.isEmpty()) {
+        return res.status(400).json({ errors: errors.array() });
+      }
       const requestId = Number(req.params.requestId);
       const request = await ops.requestService.getRequest(requestId);
       return res.status(200).json(request);
