+++ apps/funding-platform/src/request/request.service.spec.ts
@@ -219,7 +219,6 @@ describe("test RequestService class", function () {
       }
     );
     const unresolvedRequests = await requestService.getAllUnresolvedRequests();
-    console.log(unresolvedRequests);
     const unresolvedRequestsKeyedByChain =
       requestService.groupRequestsByChainId(unresolvedRequests ?? []);
     assert.equal(unresolvedRequestsKeyedByChain[1].length, 1);
